# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADIRRA API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/element/{source}/{table}/{column}/reference-data": {
      "get": {
        "tags": [
          "element"
        ],
        "summary": "Get Reference Data",
        "description": "Return reference data for a coded column (distinct_count \u2264 50).\n\nPhase 5b.2 \u2014 when ``refdata_backend='postgres'`` an *unbound* coded field returns per-code\nrows (Value / Meaning / Origin / Status) from the ``reference_code`` table, a whole-tab\n``set_badge`` (partially-approved until 100%), and a ``semantic_accepted`` gate (Submit is\ndisabled until the semantic type is Accepted).\n\n*Bound* fields (2026-08-16 redesign): every code the bound set recognises is returned\nread-only with the set's own meaning/value (``governed: true``, no per-code review of its\nown); any code the set does NOT recognise still gets the normal editable per-code treatment\n(``governed: false``) so a genuinely new/unexpected value can still be documented and\nreviewed on its own. The binding decision itself has its own submit/approve status,\nreported as ``binding_status``/``binding_submitted_at``/etc.",
        "operationId": "get_reference_data_element__source___table___column__reference_data_get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table"
            }
          },
          {
            "name": "column",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Column"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "element"
        ],
        "summary": "Update Reference Data",
        "description": "Update reference data meanings and/or status for a coded column.",
        "operationId": "update_reference_data_element__source___table___column__reference_data_patch",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table"
            }
          },
          {
            "name": "column",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Column"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schema"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceDataUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reference-data": {
      "get": {
        "tags": [
          "reference-data"
        ],
        "summary": "Get Reference Dataspace",
        "description": "Return the published reference-code register grouped by source, schema, table.\n\nEvery field is a coded column with at least one PUBLISHED code (``in_review`` or\n``approved``) from the per-code ``reference_code`` store. The field-level status\nrolls up to ``approved`` only when every published code is approved.",
        "operationId": "get_reference_dataspace_reference_data_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schema"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "semantic_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semantic Type"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Reference Dataspace Reference Data Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reference-sets": {
      "get": {
        "tags": [
          "reference-sets"
        ],
        "summary": "List Reference Sets",
        "description": "List all governed reference sets (without full entries).",
        "operationId": "list_reference_sets_reference_sets_get",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Reference Sets Reference Sets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reference-sets/{set_id}": {
      "get": {
        "tags": [
          "reference-sets"
        ],
        "summary": "Get Reference Set",
        "description": "Return one reference set including its full entry list.",
        "operationId": "get_reference_set_reference_sets__set_id__get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Reference Set Reference Sets  Set Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReferenceDataUpdateRequest": {
        "properties": {
          "meanings": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meanings"
          },
          "values": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "bound_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bound Set Id"
          },
          "unbind": {
            "type": "boolean",
            "title": "Unbind",
            "default": false
          }
        },
        "type": "object",
        "title": "ReferenceDataUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}